Changed int to size_t

+++ libgraph/include/katana/analytics/k_shortest_paths/ksssp.h
@@ -41,8 +41,8 @@ public:
 /// parameter can be specified, but have reasonable defaults.
 KATANA_EXPORT Result<void> Ksssp(
     katana::PropertyGraph* pg, const std::string& edge_weight_property_name,
-    uint32_t start_node, uint32_t report_node, katana::TxnContext* txn_ctx,
-    AlgoReachability algo_reachability, uint32_t num_paths,
+    size_t start_node, size_t report_node, katana::TxnContext* txn_ctx,
+    AlgoReachability algo_reachability, size_t num_paths,
     const bool& is_symmetric, kSsspPlan plan = {});
 }  // namespace katana::analytics
